add them directly on the ViewAnimator, and update README.md
