+++ sample/src/main/java/com/github/florent37/sample/viewanimator/MainActivity.java
@@ -8,7 +8,6 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import com.github.florent37.viewanimator.AnimationListener;
-import com.github.florent37.viewanimator.EnhancedViewAnimator;
 import com.github.florent37.viewanimator.ViewAnimator;
 
 public class MainActivity extends AppCompatActivity {
@@ -48,14 +47,14 @@ public class MainActivity extends AppCompatActivity {
         final View btnWave = findViewById(R.id.wave);
         btnWave.setOnClickListener(new View.OnClickListener() {
             @Override public void onClick(View v) {
-                EnhancedViewAnimator.animate(btnWave).wave().duration(3000).start();
+                ViewAnimator.animate(btnWave).wave().duration(3000).start();
             }
         });
 
         final View btnShake = findViewById(R.id.shake);
         btnShake.setOnClickListener(new View.OnClickListener() {
             @Override public void onClick(View v) {
-                EnhancedViewAnimator.animate(btnShake).shake().duration(3000).start();
+                ViewAnimator.animate(btnShake).shake().duration(3000).start();
             }
         });
     }
